requirements.txt
sanic
sanic-ext
ujson
requests
pymysql
cryptography
pyjwt
